Hotfix for newsfeed
The main BBC feed is using jsonp now, which isn't useful as it cannot be
parsed as JSON. This commit removes the call to that jsonp API, and
moves the setTimeout to no longer depend on the BBC News API call being
successful.

Hotfix until we make this module not awful.