Add onunload module function
This commit adds an onunload module method in the same fashion as the
existing onload method. These aren't events.

This method will be run shortly before a module is (re)loaded if the
module was already loaded, in order to allow module authors to
gracefully tear down things they set up in the onload function, such as
timers.

It doesn't run on quit.

Fixes #152.