Move UDP plugin into UDP module
The UDP stuff was previously in two separate files because we didn't
want a !reload to spin up a new socket, but now we have an onunload
event we can merge them.

This has the benefit of allowing changes to the UDP server code to take
effect without requiring a full bot restart.