Report all tweeted tweets to a channel
As part of good governance, we need to keep an eye on automated tweets
emanating from the bot, so that we can intervene if something goes
horribly wrong.

This commit adds functionality, enabled by default, where tweets are
reported to a central channel so that humans have a chance to
retroactively moderate any problematic tweets.

Ideally we will eventually add a delete tweet command to aid this.