Allow commands to opt into message events
Messages sent by the bot don't usually trigger a message event.
This commit adds a command property that, if set, causes the response of
that command to be parsed through the message event handlers, mostly.

An alternative would be to leverage the selfMessage event.

BREAKING CHANGE:

This commit removes the possibility of non-functions being passed as
commands. Previously you could just pass a string through and it would
work. None of our commands actually did that, so I've removed that to
simplify the code slightly.