Don't fail if trying to shorten a bitly link
If you try to generate a bit.ly link for a bit.ly link, the bit.ly API
doesn't like it and will return a 500.

This meant that the URL event for generating short URLs and titles will
fail, meaning that no bit.ly URL would ever get either an ofn or a title
output to channel.

This commit changes that behaviour in a somewhat hacky way so that the
HTTP 500 will be ignored and the URL suppressed in output if the URL is
already a bit.ly link.

Other 500s will fail in the usual manner.