Intercept HTTP error to add shorter message
By default the error thrown when there's a HTTP error (e.g. 404) has the
response body as its message. This isn't appropriate for the petitions
website, as it'll send a tonne of HTML at us instead of a nice JSON
response, which gets echoed into channel with a gazillion lines of text.

This commit checks to see whether a thrown error is a HTTP error, and if
so overwrites the message to a more human friendly one.