404 sensibly in the petition code (#347)
* Intercept HTTP error to add shorter message

By default the error thrown when there's a HTTP error (e.g. 404) has the
response body as its message. This isn't appropriate for the petitions
website, as it'll send a tonne of HTML at us instead of a nice JSON
response, which gets echoed into channel with a gazillion lines of text.

This commit checks to see whether a thrown error is a HTTP error, and if
so overwrites the message to a more human friendly one.

* Simplify request code to use options

This commit simplifies the request to use various library options, so
the JSON decode takes place elsewhere. We also add a user agent so that
@pixeltrix can contact us if we go dangerously haywire.

* Remove giant try/catch

Code to echo an error message is now elsewhere, which means we can get
rid of giant try/catch blocks solely to output a message to the user.

This commit removes that giant try/catch and only sticks the try around
the bit of code we're actually explicitly catching.