Tidy object properties assignment / reference
StandardJS 13 adds some new rules to disallow quotes around object
literal property names that are not strictly required.

This commit removes quotes from places that we don't need them anymore,
and moves things to dot-notation where possible.

We maintain quotes where we're abusing objects to maintain a list of
pairs of strings to test with; these are manually opted out of the
linter using eslint comments.