Tweet source name if no URL (#411)
This commit adds a check on the URL parameter to ensure it is actually a
URL; if it isn't, it won't tweet it. Reuters have previously put prose
in the "url" field, which means that they get tweeted out without being
owoified.

This commit also adds functionality to tweet out the news source if
there's no URL yet.